requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
future==0.18.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
oauthlib==3.2.2
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
types-requests==2.28.11.5
types-six==1.16.21.4
types-urllib3==1.26.25.4
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13